requirements.txt
asgiref
Django==4.0.1
djangorestframework
djongo==1.3.6
dnspython
pymongo==3.12.1
drf-yasg==1.21.7
pytz
sqlparse
celery
redis
django-celery-beat